requirements.txt
psycopg2-binary==2.9.9
fastapi==0.110.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
retry==0.9.2
uvicorn==0.29.0
pydantic==2.7.0
